Update spring boot to 2.3.3.RELEASE
This is part of an effort to synchronize the spring-boot and spring versions across our repositories.

+++ src/main/java/org/cbioportal/session_service/domain/internal/SessionRepositoryCustom.java
@@ -52,7 +52,7 @@ public interface SessionRepositoryCustom {
 
     List<Session> findBySourceAndType(String source, SessionType type);
 
-    int deleteBySourceAndTypeAndId(String source, SessionType type, String id);
+    long deleteBySourceAndTypeAndId(String source, SessionType type, String id);
 
     List<Session> findBySourceAndTypeAndQuery(String source, SessionType type, String query);
